JTabbedPane added so the user can easily navigate between opened files.
Changes:
The "new file" button has functionality now. The CTRL+N keyboard shortcut 
has functionality now. When they are clicked/pressed, a new tab will open
with the title saying "new title" and it will automatically request
focus. The tabs on the JTabbedPane now have close buttons. When opening
a file, a new tab will be opened and it's title will be the title of the
document the user selected for opening. When saving a document, the title
of the active tab will be set to the file name the user specifies.

Other Major Changes:
A new feature that deals with exception/error reporting has been added.
The ErrorReport class will be used in the try-catch block. The exception
is to be passed to the ErrorReport's createErrorReport() method. This
exception's stack trace will be written to a file in the "ErrorReporting"
directory with the extension of ".err".

A status bar has been added at the bottom, but will need better 
implementation.

Issues:
Font changes need to be implemented (by Sunday).
Tabbed Pane needs to be tested.
Plain text files need to be able to be saved.
Saving a file needs to be more user friendly (better implementation).
Opening a file needs to be tested more (better implementation).
	- Prevent the same file from being opened more than one time.
Status bar needs to be better implemented.

Upcoming events:
Code review and documentation (Saturday).
	- Code will be reviewed and properly formatted.
	- Comments will be reviewed and updated.
	- Project status will be documented.
Code reviews will happen every Saturday.